Feat(#42): 내기 게시글 참여 버튼 상태 반영, 취소 버튼 css 변경, 호스트 참여 취소 불가능하게 수정

+++ src/components/modal/DetailModal.jsx
@@ -40,6 +40,7 @@ function DetailModal({ postId, show, onHide }) {
 
   const handleJoinClick = () => {
     if (isParticipating) {
+      if (post.name === loggedIn.name) return window.alert("글 작성자는 취소할 수 없습니다. 😭");
       // 참여 취소 요청
       fetch(`http://localhost:3000/api/board/party/${postId}`, {
         method: 'DELETE',
@@ -200,7 +201,8 @@ function DetailModal({ postId, show, onHide }) {
     display: 'block',
     margin: '20px auto',
     padding: '7px 15px',
-    background: currentParticipants === maxParticipants ? '#FFFFFF' : '#022DA6',
+    // background: currentParticipants === maxParticipants ? '#FFFFFF' : '#022DA6',
+    background: currentParticipants >= maxParticipants ? '#FFFFFF' : isParticipating ? '#E24444' : '#022DA6',
     borderRadius: '10px',
     fontFamily: 'Jalnan, sans-serif',
     fontSize: '12px',
@@ -278,10 +280,16 @@ function DetailModal({ postId, show, onHide }) {
           onClick={handleJoinClick}
           disabled={currentParticipants >= maxParticipants}
         >
+            {/* {currentParticipants >= maxParticipants
+              ? '마감됨'
+              : isParticipating
+              ? '참여취소'
+              : '참여하기'} */}
+
             {/* 1. 마감된 경우(post.isEnd == true): 마감됨
             2. 마감되지 않음(post.isEnd == false), 내가 포함됨: 참여취소
             3. 마감되지 않음(post.isEnd == false), 내가 포함되지 않음: 참여하기 */}
-            {post.isEnd ? '마감됨' : isParticipating ? '참여취소' : '참여하기'}
+            {currentParticipants >= maxParticipants ? '마감됨' : isParticipating ? '참여취소' : '참여하기'}
         </button>
         <p style={participantsStyle}>
           현재 참여 인원:{' '}
